feat: implement read_network and persist_network
For the NETWORK table went with wallet_name as the key to allow for
one db storing multiple wallets. Also took wallet_name as &str since
would need to pass it on to various persist functions(considering
adding this as a field in Store), while for network went with String as
it seemed easier to convert bitcoin::Network to.

The redb::ReadTransaction::open_table function throws error if table
hasn't been created already so planning to introduce an init_tables
function in another commit. Adopted bdk_sqlx style and chose to panic
rather than return an error if network is parsed incorrectly. In case of
network not being persisted already chose to return error instead since
this might happen easily if user is not careful IMO.

Also chose to not differentiate between read_txn and write_txn in naming
following bdk_sqlx, rusqlite_impl etc. could change that if necessary