feat: create tables before reading values
Added this since `redb::ReadTransaction::open_table` returns an
error if the table hasn't been created earlier. Although this does not
solve the problem when the read_network function goes on to read
the network. Feel like having this function would be better than giving
a redb error to the user but we can remove this in case its unnecessary.