feat: implement keychain persistence
persist_keychains:
Followed bdk_sqlx for keychains table. Initially my approach was to
replace keychainkind with a u64 but ig having labels makes bdk/issues/57
easier.
Chose (wallet_name, label) as key because it would make it easier to
retrieve a particular descriptor if we choose to do so in the future.

read_keychains:
Added a num_keychains field to ensure that expected number of keychains
have been persisted and to handle the single descriptor case.