refactor!: make local_chain persistence modular
So that we donot need to persist the whole local_chain every time
local_chain::ChangeSet has non trivial entries. Persisting the whole
local_chain may have performance costs when the number of blocks is
huge.

An alternate approach could have been to use MultimapTable but during
persistence if a checkpoint is being removed that would mean going over
all blocks synced by a wallet! In the worst case this would mean
quadratic complexity in number of blocks.

Introduced chain_height as an Option in case chain_height is not known
before loading.