refactor: make last_revealed persistence modular
Instead of persisting the whole last_revealed blob each time indexer
has non trivial entries we now just have to persist the changes for the
descriptors involved.

Another approach was using MultimapTable but that became a bit painful
to implement. To make loading faster if we know the descriptors of
interest before loading the `desc_ids` parameter can be used.
ToDo:
A wallet would know descriptors before hand due to `read_keychains` so
it is possible to refactor `persist_changeset` to use this power.