feat: persist anchors per wallet
Since an anchor can correspond to multiple txns and a txn can correspond
to multiple anchors (and taking inspiration from rusqlite_impl) took
(Txid, BlockId) to be key and the anchor metadata to be value. ToDo:
implement for generic Anchor case.
Also added derive PartialEq and Eq for wrappers so that keys donot get
duplicated.
Included a test to check persistence of anchors.